+++ lualib/skynet/remotedebug.lua
@@ -1,6 +1,6 @@
 local skynet = require "skynet"
 local debugchannel = require "debugchannel"
-local socket = require "socket"
+local socketdriver = require "socketdriver"
 local injectrun = require "skynet.injectcode"
 local table = table
 local debug = debug
@@ -56,7 +56,7 @@ local function gen_print(fd)
 			tmp[i] = tostring(tmp[i])
 		end
 		table.insert(tmp, "\n")
-		socket.write(fd, table.concat(tmp, "\t"))
+		socketdriver.send(fd, table.concat(tmp, "\t"))
 	end
 end
 
@@ -140,9 +140,12 @@ end
 local function watch_proto(protoname, cond)
 	local proto = assert(replace_upvalue(skynet.register_protocol, "proto"), "Can't find proto table")
 	local p = proto[protoname]
-	local dispatch = p.dispatch_origin or p.dispatch
-	if p == nil or dispatch == nil then
+	if p == nil then
 		return "No " .. protoname
+	end
+	local dispatch = p.dispatch_origin or p.dispatch
+	if dispatch == nil then
+		return "No dispatch"
 	end
 	p.dispatch_origin = dispatch
 	p.dispatch = function(...)
@@ -208,7 +211,7 @@ local function hook_dispatch(dispatcher, resp, fd, channel)
 	local function debug_hook()
 		while true do
 			if newline then
-				socket.write(fd, prompt)
+				socketdriver.send(fd, prompt)
 				newline = false
 			end
 			local cmd = channel:read()
@@ -244,7 +247,9 @@ local function hook_dispatch(dispatcher, resp, fd, channel)
 	func = replace_upvalue(dispatcher, HOOK_FUNC, hook)
 	if func then
 		local function idle()
-			skynet.timeout(10,idle)	-- idle every 0.1s
+			if raw_dispatcher then
+			    skynet.timeout(10,idle)	-- idle every 0.1s
+			end
 		end
 		skynet.timeout(0, idle)
 	end