+++ service/debug_console.lua
@@ -4,6 +4,8 @@ local core = require "skynet.core"
 local socket = require "socket"
 local snax = require "snax"
 local memory = require "memory"
+local httpd = require "http.httpd"
+local sockethelper = require "http.sockethelper"
 
 local port = tonumber(...)
 local COMMAND = {}
@@ -52,13 +54,10 @@ end
 local function docmd(cmdline, print, fd)
 	local split = split_cmdline(cmdline)
 	local command = split[1]
-	if command == "debug" then
-		table.insert(split, fd)
-	end
 	local cmd = COMMAND[command]
 	local ok, list
 	if cmd then
-		ok, list = pcall(cmd, select(2,table.unpack(split)))
+		ok, list = pcall(cmd, fd, select(2,table.unpack(split)))
 	else
 		print("Invalid command, type help for command list")
 	end
@@ -79,18 +78,28 @@ local function docmd(cmdline, print, fd)
 end
 
 local function console_main_loop(stdin, print)
-	socket.lock(stdin)
 	print("Welcome to skynet console")
-	while true do
-		local cmdline = socket.readline(stdin, "\n")
-		if not cmdline then
-			break
+	skynet.error(stdin, "connected")
+	pcall(function()
+		while true do
+			local cmdline = socket.readline(stdin, "\n")
+			if not cmdline then
+				break
+			end
+			if cmdline:sub(1,4) == "GET " then
+				-- http
+				local code, url = httpd.read_request(sockethelper.readfunc(stdin, cmdline.. "\n"), 8192)
+				local cmdline = url:sub(2):gsub("/"," ")
+				docmd(cmdline, print, stdin)
+				break
+			end
+			if cmdline ~= "" then
+				docmd(cmdline, print, stdin)
+			end
 		end
-		if cmdline ~= "" then
-			docmd(cmdline, print, stdin)
-		end
-	end
-	socket.unlock(stdin)
+	end)
+	skynet.error(stdin, "disconnected")
+	socket.close(stdin)
 end
 
 skynet.start(function()
@@ -115,7 +124,7 @@ function COMMAND.help()
 		help = "This help message",
 		list = "List all the service",
 		stat = "Dump all stats",
-		info = "Info address : get service infomation",
+		info = "info address : get service infomation",
 		exit = "exit address : kill a lua service",
 		kill = "kill address : kill service",
 		mem = "mem : show memory status",
@@ -133,6 +142,7 @@ function COMMAND.help()
 		signal = "signal address sig",
 		cmem = "Show C memory info",
 		shrtbl = "Show shared short string table info",
+		ping = "ping address",
 	}
 end
 
@@ -140,25 +150,33 @@ function COMMAND.clearcache()
 	codecache.clear()
 end
 
-function COMMAND.start(...)
+function COMMAND.start(fd, ...)
 	local ok, addr = pcall(skynet.newservice, ...)
 	if ok then
-		return { [skynet.address(addr)] = ... }
+		if addr then
+			return { [skynet.address(addr)] = ... }
+		else
+			return "Exit"
+		end
 	else
 		return "Failed"
 	end
 end
 
-function COMMAND.log(...)
+function COMMAND.log(fd, ...)
 	local ok, addr = pcall(skynet.call, ".launcher", "lua", "LOGLAUNCH", "snlua", ...)
 	if ok then
-		return { [skynet.address(addr)] = ... }
+		if addr then
+			return { [skynet.address(addr)] = ... }
+		else
+			return "Failed"
+		end
 	else
 		return "Failed"
 	end
 end
 
-function COMMAND.snax(...)
+function COMMAND.snax(fd, ...)
 	local ok, s = pcall(snax.newservice, ...)
 	if ok then
 		local addr = s.handle
@@ -191,7 +209,7 @@ function COMMAND.mem()
 	return skynet.call(".launcher", "lua", "MEM")
 end
 
-function COMMAND.kill(address)
+function COMMAND.kill(fd, address)
 	return skynet.call(".launcher", "lua", "KILL", address)
 end
 
@@ -199,11 +217,11 @@ function COMMAND.gc()
 	return skynet.call(".launcher", "lua", "GC")
 end
 
-function COMMAND.exit(address)
+function COMMAND.exit(fd, address)
 	skynet.send(adjust_address(address), "debug", "EXIT")
 end
 
-function COMMAND.inject(address, filename)
+function COMMAND.inject(fd, address, filename)
 	address = adjust_address(address)
 	local f = io.open(filename, "rb")
 	if not f then
@@ -214,24 +232,24 @@ function COMMAND.inject(address, filename)
 	return skynet.call(address, "debug", "RUN", source, filename)
 end
 
-function COMMAND.task(address)
+function COMMAND.task(fd, address)
 	address = adjust_address(address)
 	return skynet.call(address,"debug","TASK")
 end
 
-function COMMAND.info(address)
+function COMMAND.info(fd, address, ...)
 	address = adjust_address(address)
-	return skynet.call(address,"debug","INFO")
+	return skynet.call(address,"debug","INFO", ...)
 end
 
-function COMMAND.debug(address, fd)
+function COMMAND.debug(fd, address)
 	address = adjust_address(address)
 	local agent = skynet.newservice "debug_agent"
 	local stop
 	skynet.fork(function()
 		repeat
 			local cmdline = socket.readline(fd, "\n")
-            cmdline = cmdline:gsub("(.*)\r$", "%1")
+			cmdline = cmdline and cmdline:gsub("(.*)\r$", "%1")
 			if not cmdline then
 				skynet.send(agent, "lua", "cmd", "cont")
 				break
@@ -243,17 +261,17 @@ function COMMAND.debug(address, fd)
 	stop = true
 end
 
-function COMMAND.logon(address)
+function COMMAND.logon(fd, address)
 	address = adjust_address(address)
 	core.command("LOGON", skynet.address(address))
 end
 
-function COMMAND.logoff(address)
+function COMMAND.logoff(fd, address)
 	address = adjust_address(address)
 	core.command("LOGOFF", skynet.address(address))
 end
 
-function COMMAND.signal(address, sig)
+function COMMAND.signal(fd, address, sig)
 	address = skynet.address(adjust_address(address))
 	if sig then
 		core.command("SIGNAL", string.format("%s %d",address,sig))
@@ -276,4 +294,10 @@ function COMMAND.shrtbl()
 	return { n = n, total = total, longest = longest, space = space }
 end
 
-
+function COMMAND.ping(fd, address)
+	address = adjust_address(address)
+	local ti = skynet.now()
+	skynet.call(address, "debug", "PING")
+	ti = skynet.now() - ti
+	return tostring(ti)
+end