+++ skynet-src/malloc_hook.h
@@ -11,6 +11,7 @@ extern size_t mallctl_int64(const char* name, size_t* newval);
 extern int    mallctl_opt(const char* name, int* newval);
 extern void   dump_c_mem(void);
 extern int    dump_mem_lua(lua_State *L);
+extern size_t malloc_current_memory(void);
 
 #endif /* SKYNET_MALLOC_HOOK_H */